Process sign-in requests
There is some misleading information regarding the Sign-in request process:

"The Authentication Agent forwards the result back to Azure AD STS over an outbound mutually authenticated HTTPS channel over port 443. Mutual authentication uses the certificate previously issued to the Authentication Agent during registration." 

It's not clear that this outbound call is done through the already established connection to the servicebus endpoint or a through a new connection.

Since this already created some confusion with customers, I would like to request an update in this documentation to include these details.

Thank you for your help.